fix: try fix clippy issure

+++ chat/chat_server/src/models/messages.rs
@@ -98,7 +98,7 @@ impl AppState {
                 "#,
             )
             .bind(chat_id as i64)
-            .bind(other_user_id as i64)
+            .bind(other_user_id)
             .bind(reply)
             .fetch_one(&self.pool)
             .await?;
